refactor!: Make `Typed::Result` sealed
We only ever want there to be two inheritors of `Typed::Result`, `Success` and `Failure`. We can express this in Sorbet with the `sealed!` helper. In order to use this, we must move the definitions of `Success` and `Failure` into the same file as `Result`. See https://sorbet.org/docs/sealed for more info.
